+++ buession-redis/src/main/java/com/buession/redis/client/jedis/operations/JedisClusterHashOperations.java
@@ -441,7 +441,7 @@ public final class JedisClusterHashOperations extends AbstractHashOperations<Jed
 	}
 
 	@Override
-	public List<String> hRandField(final String key, final long count) {
+	public List<String> hRandField(final String key, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("count", count);
 
 		if(isPipeline()){
@@ -460,7 +460,7 @@ public final class JedisClusterHashOperations extends AbstractHashOperations<Jed
 	}
 
 	@Override
-	public List<byte[]> hRandField(final byte[] key, final long count) {
+	public List<byte[]> hRandField(final byte[] key, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("count", count);
 
 		if(isPipeline()){
@@ -479,7 +479,7 @@ public final class JedisClusterHashOperations extends AbstractHashOperations<Jed
 	}
 
 	@Override
-	public List<KeyValue<String, String>> hRandFieldWithValues(final String key, final long count) {
+	public List<KeyValue<String, String>> hRandFieldWithValues(final String key, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("count", count);
 		final ListConverter<Map.Entry<String, String>, KeyValue<String, String>> converter =
 				new ListConverter<>(new MapEntryKeyValueConverter<>((k)->k, (v)->v));
@@ -500,7 +500,7 @@ public final class JedisClusterHashOperations extends AbstractHashOperations<Jed
 	}
 
 	@Override
-	public List<KeyValue<byte[], byte[]>> hRandFieldWithValues(final byte[] key, final long count) {
+	public List<KeyValue<byte[], byte[]>> hRandFieldWithValues(final byte[] key, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("count", count);
 		final ListConverter<Map.Entry<byte[], byte[]>, KeyValue<byte[], byte[]>> converter =
 				new ListConverter<>(new MapEntryKeyValueConverter<>((k)->k, (v)->v));
@@ -579,7 +579,7 @@ public final class JedisClusterHashOperations extends AbstractHashOperations<Jed
 	}
 
 	@Override
-	public ScanResult<Map<String, String>> hScan(final String key, final String cursor, final long count) {
+	public ScanResult<Map<String, String>> hScan(final String key, final String cursor, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("cursor", cursor).put("count", count);
 		final ScanParams scanParams = new JedisScanParams(count);
 
@@ -587,7 +587,7 @@ public final class JedisClusterHashOperations extends AbstractHashOperations<Jed
 	}
 
 	@Override
-	public ScanResult<Map<byte[], byte[]>> hScan(final byte[] key, final byte[] cursor, final long count) {
+	public ScanResult<Map<byte[], byte[]>> hScan(final byte[] key, final byte[] cursor, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("cursor", cursor).put("count", count);
 		final ScanParams scanParams = new JedisScanParams(count);
 
@@ -596,7 +596,7 @@ public final class JedisClusterHashOperations extends AbstractHashOperations<Jed
 
 	@Override
 	public ScanResult<Map<String, String>> hScan(final String key, final String cursor, final String pattern,
-												 final long count) {
+												 final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("cursor", cursor).put("pattern", pattern);
 		final ScanParams scanParams = new JedisScanParams(pattern, count);
 
@@ -605,7 +605,7 @@ public final class JedisClusterHashOperations extends AbstractHashOperations<Jed
 
 	@Override
 	public ScanResult<Map<byte[], byte[]>> hScan(final byte[] key, final byte[] cursor, final byte[] pattern,
-												 final long count) {
+												 final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("cursor", cursor).put("pattern", pattern)
 				.put("count", count);
 		final ScanParams scanParams = new JedisScanParams(pattern, count);
