+++ buession-redis/src/main/java/com/buession/redis/client/jedis/operations/JedisClusterSetOperations.java
@@ -437,7 +437,7 @@ public final class JedisClusterSetOperations extends AbstractSetOperations<Jedis
 	}
 
 	@Override
-	public Set<String> sPop(final String key, final long count) {
+	public Set<String> sPop(final String key, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("count", count);
 
 		if(isPipeline()){
@@ -454,7 +454,7 @@ public final class JedisClusterSetOperations extends AbstractSetOperations<Jedis
 	}
 
 	@Override
-	public Set<byte[]> sPop(final byte[] key, final long count) {
+	public Set<byte[]> sPop(final byte[] key, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("count", count);
 
 		if(isPipeline()){
@@ -507,7 +507,7 @@ public final class JedisClusterSetOperations extends AbstractSetOperations<Jedis
 	}
 
 	@Override
-	public List<String> sRandMember(final String key, final long count) {
+	public List<String> sRandMember(final String key, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("count", count);
 
 		if(isPipeline()){
@@ -526,7 +526,7 @@ public final class JedisClusterSetOperations extends AbstractSetOperations<Jedis
 	}
 
 	@Override
-	public List<byte[]> sRandMember(final byte[] key, final long count) {
+	public List<byte[]> sRandMember(final byte[] key, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("count", count);
 
 		if(isPipeline()){
@@ -639,7 +639,7 @@ public final class JedisClusterSetOperations extends AbstractSetOperations<Jedis
 	}
 
 	@Override
-	public ScanResult<List<String>> sScan(final String key, final String cursor, final long count) {
+	public ScanResult<List<String>> sScan(final String key, final String cursor, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("cursor", cursor).put("count", count);
 		final ScanParams scanParams = new JedisScanParams(count);
 
@@ -647,7 +647,7 @@ public final class JedisClusterSetOperations extends AbstractSetOperations<Jedis
 	}
 
 	@Override
-	public ScanResult<List<byte[]>> sScan(final byte[] key, final byte[] cursor, final long count) {
+	public ScanResult<List<byte[]>> sScan(final byte[] key, final byte[] cursor, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("cursor", cursor).put("count", count);
 		final ScanParams scanParams = new JedisScanParams(count);
 
@@ -656,7 +656,7 @@ public final class JedisClusterSetOperations extends AbstractSetOperations<Jedis
 
 	@Override
 	public ScanResult<List<String>> sScan(final String key, final String cursor, final String pattern,
-										  final long count) {
+										  final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("cursor", cursor).put("pattern", pattern)
 				.put("count", count);
 		final ScanParams scanParams = new JedisScanParams(pattern, count);
@@ -666,7 +666,7 @@ public final class JedisClusterSetOperations extends AbstractSetOperations<Jedis
 
 	@Override
 	public ScanResult<List<byte[]>> sScan(final byte[] key, final byte[] cursor, final byte[] pattern,
-										  final long count) {
+										  final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("cursor", cursor).put("pattern", pattern)
 				.put("count", count);
 		final ScanParams scanParams = new JedisScanParams(pattern, count);
