+++ buession-redis/src/main/java/com/buession/redis/client/jedis/operations/JedisHashOperations.java
@@ -424,7 +424,7 @@ public final class JedisHashOperations extends AbstractHashOperations<JedisStand
 	}
 
 	@Override
-	public List<String> hRandField(final String key, final long count) {
+	public List<String> hRandField(final String key, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("count", count);
 
 		if(isPipeline()){
@@ -442,7 +442,7 @@ public final class JedisHashOperations extends AbstractHashOperations<JedisStand
 	}
 
 	@Override
-	public List<byte[]> hRandField(final byte[] key, final long count) {
+	public List<byte[]> hRandField(final byte[] key, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("count", count);
 
 		if(isPipeline()){
@@ -460,7 +460,7 @@ public final class JedisHashOperations extends AbstractHashOperations<JedisStand
 	}
 
 	@Override
-	public List<KeyValue<String, String>> hRandFieldWithValues(final String key, final long count) {
+	public List<KeyValue<String, String>> hRandFieldWithValues(final String key, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("count", count);
 		final ListConverter<Map.Entry<String, String>, KeyValue<String, String>> converter =
 				new ListConverter<>(new MapEntryKeyValueConverter<>((k)->k, (v)->v));
@@ -481,7 +481,7 @@ public final class JedisHashOperations extends AbstractHashOperations<JedisStand
 	}
 
 	@Override
-	public List<KeyValue<byte[], byte[]>> hRandFieldWithValues(final byte[] key, final long count) {
+	public List<KeyValue<byte[], byte[]>> hRandFieldWithValues(final byte[] key, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("count", count);
 		final ListConverter<Map.Entry<byte[], byte[]>, KeyValue<byte[], byte[]>> converter =
 				new ListConverter<>(new MapEntryKeyValueConverter<>((k)->k, (v)->v));
@@ -560,7 +560,7 @@ public final class JedisHashOperations extends AbstractHashOperations<JedisStand
 	}
 
 	@Override
-	public ScanResult<Map<String, String>> hScan(final String key, final String cursor, final long count) {
+	public ScanResult<Map<String, String>> hScan(final String key, final String cursor, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("cursor", cursor).put("count", count);
 		final ScanParams scanParams = new JedisScanParams(count);
 
@@ -568,7 +568,7 @@ public final class JedisHashOperations extends AbstractHashOperations<JedisStand
 	}
 
 	@Override
-	public ScanResult<Map<byte[], byte[]>> hScan(final byte[] key, final byte[] cursor, final long count) {
+	public ScanResult<Map<byte[], byte[]>> hScan(final byte[] key, final byte[] cursor, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("cursor", cursor).put("count", count);
 		final ScanParams scanParams = new JedisScanParams(count);
 
@@ -577,7 +577,7 @@ public final class JedisHashOperations extends AbstractHashOperations<JedisStand
 
 	@Override
 	public ScanResult<Map<String, String>> hScan(final String key, final String cursor, final String pattern,
-												 final long count) {
+												 final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("cursor", cursor).put("pattern", pattern);
 		final ScanParams scanParams = new JedisScanParams(pattern, count);
 
@@ -586,7 +586,7 @@ public final class JedisHashOperations extends AbstractHashOperations<JedisStand
 
 	@Override
 	public ScanResult<Map<byte[], byte[]>> hScan(final byte[] key, final byte[] cursor, final byte[] pattern,
-												 final long count) {
+												 final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("cursor", cursor).put("pattern", pattern)
 				.put("count", count);
 		final ScanParams scanParams = new JedisScanParams(pattern, count);
