+++ buession-redis/src/main/java/com/buession/redis/client/jedis/operations/JedisKeyOperations.java
@@ -958,7 +958,7 @@ public final class JedisKeyOperations extends AbstractKeyOperations<JedisStandal
 	}
 
 	@Override
-	public ScanResult<List<String>> scan(final String cursor, final long count) {
+	public ScanResult<List<String>> scan(final String cursor, final int count) {
 		final CommandArguments args = CommandArguments.create("cursor", cursor).put("count", count);
 		final ScanParams scanParams = new JedisScanParams(count);
 
@@ -966,7 +966,7 @@ public final class JedisKeyOperations extends AbstractKeyOperations<JedisStandal
 	}
 
 	@Override
-	public ScanResult<List<byte[]>> scan(final byte[] cursor, final long count) {
+	public ScanResult<List<byte[]>> scan(final byte[] cursor, final int count) {
 		final CommandArguments args = CommandArguments.create("cursor", cursor).put("count", count);
 		final ScanParams scanParams = new JedisScanParams(count);
 
@@ -974,7 +974,7 @@ public final class JedisKeyOperations extends AbstractKeyOperations<JedisStandal
 	}
 
 	@Override
-	public ScanResult<List<String>> scan(final String cursor, final String pattern, final long count) {
+	public ScanResult<List<String>> scan(final String cursor, final String pattern, final int count) {
 		final CommandArguments args = CommandArguments.create("cursor", cursor).put("pattern", pattern)
 				.put("count", count);
 		final ScanParams scanParams = new JedisScanParams(pattern, count);
@@ -983,7 +983,7 @@ public final class JedisKeyOperations extends AbstractKeyOperations<JedisStandal
 	}
 
 	@Override
-	public ScanResult<List<byte[]>> scan(final byte[] cursor, final byte[] pattern, final long count) {
+	public ScanResult<List<byte[]>> scan(final byte[] cursor, final byte[] pattern, final int count) {
 		final CommandArguments args = CommandArguments.create("cursor", cursor).put("pattern", pattern)
 				.put("count", count);
 		final ScanParams scanParams = new JedisScanParams(pattern, count);
