+++ buession-redis/src/main/java/com/buession/redis/client/jedis/operations/JedisListOperations.java
@@ -303,7 +303,7 @@ public final class JedisListOperations extends AbstractListOperations<JedisStand
 	}
 
 	@Override
-	public List<Long> lPos(final String key, final String element, final LPosArgument lPosArgument, final long count) {
+	public List<Long> lPos(final String key, final String element, final LPosArgument lPosArgument, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("lPosArgument", lPosArgument)
 				.put("count", count);
 		final LPosParams lPosParams = JedisLPosParams.from(lPosArgument);
@@ -324,7 +324,7 @@ public final class JedisListOperations extends AbstractListOperations<JedisStand
 	}
 
 	@Override
-	public List<Long> lPos(final byte[] key, final byte[] element, final LPosArgument lPosArgument, final long count) {
+	public List<Long> lPos(final byte[] key, final byte[] element, final LPosArgument lPosArgument, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("lPosArgument", lPosArgument)
 				.put("count", count);
 		final LPosParams lPosParams = JedisLPosParams.from(lPosArgument);
@@ -345,7 +345,7 @@ public final class JedisListOperations extends AbstractListOperations<JedisStand
 	}
 
 	@Override
-	public Long lRem(final String key, final String value, final long count) {
+	public Long lRem(final String key, final String value, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("value", value).put("count", count);
 
 		if(isPipeline()){
@@ -362,7 +362,7 @@ public final class JedisListOperations extends AbstractListOperations<JedisStand
 	}
 
 	@Override
-	public Long lRem(final byte[] key, final byte[] value, final long count) {
+	public Long lRem(final byte[] key, final byte[] value, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("value", value).put("count", count);
 
 		if(isPipeline()){
