+++ buession-redis/src/main/java/com/buession/redis/client/jedis/operations/JedisSentinelKeyOperations.java
@@ -970,7 +970,7 @@ public final class JedisSentinelKeyOperations extends AbstractKeyOperations<Jedi
 	}
 
 	@Override
-	public ScanResult<List<String>> scan(final String cursor, final long count) {
+	public ScanResult<List<String>> scan(final String cursor, final int count) {
 		final CommandArguments args = CommandArguments.create("cursor", cursor).put("count", count);
 		final ScanParams scanParams = new JedisScanParams(count);
 
@@ -978,7 +978,7 @@ public final class JedisSentinelKeyOperations extends AbstractKeyOperations<Jedi
 	}
 
 	@Override
-	public ScanResult<List<byte[]>> scan(final byte[] cursor, final long count) {
+	public ScanResult<List<byte[]>> scan(final byte[] cursor, final int count) {
 		final CommandArguments args = CommandArguments.create("cursor", cursor).put("count", count);
 		final ScanParams scanParams = new JedisScanParams(count);
 
@@ -986,7 +986,7 @@ public final class JedisSentinelKeyOperations extends AbstractKeyOperations<Jedi
 	}
 
 	@Override
-	public ScanResult<List<String>> scan(final String cursor, final String pattern, final long count) {
+	public ScanResult<List<String>> scan(final String cursor, final String pattern, final int count) {
 		final CommandArguments args = CommandArguments.create("cursor", cursor).put("pattern", pattern)
 				.put("count", count);
 		final ScanParams scanParams = new JedisScanParams(pattern, count);
@@ -995,7 +995,7 @@ public final class JedisSentinelKeyOperations extends AbstractKeyOperations<Jedi
 	}
 
 	@Override
-	public ScanResult<List<byte[]>> scan(final byte[] cursor, final byte[] pattern, final long count) {
+	public ScanResult<List<byte[]>> scan(final byte[] cursor, final byte[] pattern, final int count) {
 		final CommandArguments args = CommandArguments.create("cursor", cursor).put("pattern", pattern)
 				.put("count", count);
 		final ScanParams scanParams = new JedisScanParams(pattern, count);
