+++ buession-redis/src/main/java/com/buession/redis/client/jedis/operations/JedisSentinelListOperations.java
@@ -314,7 +314,7 @@ public final class JedisSentinelListOperations extends AbstractListOperations<Je
 	}
 
 	@Override
-	public List<Long> lPos(final String key, final String element, final LPosArgument lPosArgument, final long count) {
+	public List<Long> lPos(final String key, final String element, final LPosArgument lPosArgument, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("lPosArgument", lPosArgument)
 				.put("count", count);
 		final LPosParams lPosParams = JedisLPosParams.from(lPosArgument);
@@ -335,7 +335,7 @@ public final class JedisSentinelListOperations extends AbstractListOperations<Je
 	}
 
 	@Override
-	public List<Long> lPos(final byte[] key, final byte[] element, final LPosArgument lPosArgument, final long count) {
+	public List<Long> lPos(final byte[] key, final byte[] element, final LPosArgument lPosArgument, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("lPosArgument", lPosArgument)
 				.put("count", count);
 		final LPosParams lPosParams = JedisLPosParams.from(lPosArgument);
@@ -356,7 +356,7 @@ public final class JedisSentinelListOperations extends AbstractListOperations<Je
 	}
 
 	@Override
-	public Long lRem(final String key, final String value, final long count) {
+	public Long lRem(final String key, final String value, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("value", value).put("count", count);
 
 		if(isPipeline()){
@@ -374,7 +374,7 @@ public final class JedisSentinelListOperations extends AbstractListOperations<Je
 	}
 
 	@Override
-	public Long lRem(final byte[] key, final byte[] value, final long count) {
+	public Long lRem(final byte[] key, final byte[] value, final int count) {
 		final CommandArguments args = CommandArguments.create("key", key).put("value", value).put("count", count);
 
 		if(isPipeline()){
